removed health scripts; shooting enemies might work (haven't tried)
