+++ Assets/Scripts/Enemies/Enemy.cs
@@ -2,21 +2,62 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Enemy : MonoBehaviour
+public class Enemy : MonoBehaviour, IShootable
 {
     [SerializeField] protected EnemyPathing enemyPathing;
 
-    [SerializeField] protected Health health;
+    [SerializeField] protected int maxHealth = 1;
+    public int currentHealth {get; private set;}
+
     [SerializeField] protected int attackValue = 1;
     [SerializeField] protected bool damageOnImpact;
 
+    void Start()
+    {
+        currentHealth = maxHealth;
+    }
+
     void OnCollisionEnter(Collision collision)
     {
-        // TODO: if the collision is the player's gunshot, take that kind of damage
-        // health.TakeDamage(GameManager.instance.player.gunAttackDamage);
-
         if(damageOnImpact && collision.gameObject.tag == "Player"){
-            health.TakeDamage(GameManager.instance.player.motorcycleImpactDamage);
+            TakeDamage(GameManager.instance.player.motorcycleImpactDamage);
         }
     }
+
+    public void OnShoot()
+    {
+        TakeDamage(GameManager.instance.player.gunDamage);
+    }
+
+    #region Health Stuff
+        public int GetMaxHealth()
+        {
+            return maxHealth;
+        }
+
+        public virtual void SetMaxHealth(int newMaxHealth)
+        {
+            maxHealth = newMaxHealth;
+        }
+
+        public virtual void TakeDamage(int damageValue)
+        {
+            currentHealth -= damageValue;
+
+            if(currentHealth <= 0){
+                // TODO: Do death stuff
+
+                Destroy(gameObject);
+            }
+        }
+
+        public virtual void Heal(int healValue)
+        {
+            currentHealth += healValue;
+
+            if(currentHealth > maxHealth){
+                currentHealth = maxHealth;
+            }
+        }
+    #endregion
 }
