--- Assets/Scripts/Health Stuff.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a10abe3acaf7b2f43b05abcd67f7e10a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
