--- Assets/Scripts/Health Stuff/EnemyHealth.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d38737922bbca6f4387e8c6d1dbca413
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
