--- Assets/Scripts/Health Stuff/Health.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b1d46f979d3e64a438dfc1400b226be0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
