--- Assets/Scripts/Health Stuff/PlayerHealth.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 52a5e64f3af442f45adfe67a623f1ddb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
